Update update script to update antigen plugins
Also changes execution environment from bash to zsh, which should not be
a problem as I use zsh on all workstations.
However, it could be wise to wrap updating antigen plugins to a
different script that is only run when zsh and antigen are installed,
thus making script work in bash environment as well.